Fix a bug where extend does not return an Object.
'extend', when compiled to js, was returning an array of
arrays rather than an object due to implicit return.

This broke the options sent to setCookie, because the 'expires'
attribute of the returned options object was always undefined,
so the cookie expiration field was never set as
options.expires.toGMTString().

As a result, the abbaVariant and abbaPersistComplete cookies
incorrectly had no expiration date or a received a default expiration
from the browser, rather than the 600 day expiration specified
in the code.